% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AIC}
\alias{AIC}
\title{Compute AIC score of an estimated precision matrix}
\usage{
AIC(S, theta, n, stop_overflow)
}
\arguments{
\item{S}{\eqn{p} by \eqn{p} scatter matrix of the data}

\item{theta}{estimated \eqn{p} by \eqn{p} precision matrix}

\item{n}{the number of observations}

\item{stop_overflow}{should overflow be avoided?}
}
\value{
The AIC score of the estimated precision matrix
}
\description{
This function computes the AIC score of an estimated precision matrix
}
