% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ECM_GHS}
\alias{ECM_GHS}
\title{Perform GHS with ECM}
\usage{
ECM_GHS(
  X,
  S,
  theta,
  sigma,
  Lambda_sq,
  epsilon,
  verbose,
  maxitr,
  savepath,
  save_Q,
  tau_sq,
  machine_eps,
  fix_tau = FALSE,
  stop_underflow = FALSE,
  savepath_tau = FALSE
)
}
\arguments{
\item{X}{\eqn{n} by \eqn{p} matrix of data}

\item{S}{\eqn{p} by \eqn{p} scatter matrix of the data}

\item{theta}{initial value of the \eqn{p} by \eqn{p} precision matrix}

\item{sigma}{initial value of the \eqn{p} by \eqn{p} covariance matrix}

\item{Lambda_sq}{initial value of matrix of squared local shrinkage parameters}

\item{epsilon}{tolerance for the convergence assessment}

\item{verbose}{logical indicator of printing information at each iteration}

\item{maxitr}{maximum number of iterations}

\item{savepath}{logical indicator of saving the estimator at each iteration in the ECM algorithm. Only available for p<200}

\item{save_Q}{should the value of the objective function at each step be saved?}

\item{tau_sq}{initial value of squared global shrinkage parameter. If exist_group==T, a dummy value should be provided}

\item{machine_eps}{numerical. The machine precision}

\item{fix_tau}{should tau_sq be fixed?}

\item{stop_underflow}{Should tricks to avoid underflow be used?}

\item{savepath_tau}{if fix_tau==F, should its path be saved?}
}
\value{
A List with resulting ECM estimates, and saved path and objective function convergence information if requested
}
\description{
This function performs expectation-conditional-maximation or iterated conditional mode estimation for the graphical horseshoe
}
