% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ECM_GHS_AIC}
\alias{ECM_GHS_AIC}
\title{Perform GHS with ECM, selecting tau_sq by AIC}
\usage{
ECM_GHS_AIC(
  X,
  S,
  theta,
  sigma,
  Lambda_sq,
  AIC_eps,
  tau_sq_vec,
  epsilon,
  verbose,
  maxitr,
  savepath,
  save_Q,
  tau_sq,
  machine_eps,
  stop_underflow = FALSE,
  stop_overflow = FALSE
)
}
\arguments{
\item{X}{\eqn{n} by \eqn{p} matrix of data}

\item{S}{\eqn{p} by \eqn{p} scatter matrix of the data}

\item{theta}{initial value of the \eqn{p} by \eqn{p} precision matrix}

\item{sigma}{initial value of the \eqn{p} by \eqn{p} covariance matrix}

\item{Lambda_sq}{initial value of matrix of squared local shrinkage parameters}

\item{AIC_eps}{if AIC_selection == TRUE, the convergence tolerance for the AIC convergence assessment}

\item{tau_sq_vec}{values of tau_sq to consider}

\item{epsilon}{tolerance for the convergence assessment}

\item{verbose}{logical indicator of printing information at each iteration}

\item{maxitr}{maximum number of iterations}

\item{savepath}{logical indicator of saving the estimator at each iteration in the ECM algorithm. Only available for p<200}

\item{save_Q}{should the value of the objective function at each step be saved?}

\item{tau_sq}{initial value of squared global shrinkage parameter. If exist_group==T, a dummy value should be provided}

\item{machine_eps}{numerical. The machine precision}

\item{stop_underflow}{Should tricks to avoid underflow be used?}

\item{stop_overflow}{Should tricks to avoid overflow be used?}
}
\value{
A List with resulting ECM estimates, and saved path and objective function convergence information if requested
}
\description{
This function performs expectation-conditional-maximation or iterated conditional mode estimation for the graphical horseshoe
}
