% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGHS.r
\name{fastGHS}
\alias{fastGHS}
\title{Perform GHS with ECM}
\usage{
fastGHS(
  X,
  epsilon = 1e-05,
  AIC_eps = 0.1,
  theta = NULL,
  sigma = NULL,
  Lambda_sq = NULL,
  tau_sq = NULL,
  tau_sq_min = 1e-04,
  tau_sq_stepsize = NULL,
  tau_sq_max = 20,
  stop_overflow = FALSE,
  maxitr = 10000,
  verbose = TRUE,
  savepath = FALSE,
  save_Q = F,
  fix_tau = FALSE,
  AIC_selection = FALSE,
  stop_underflow = FALSE,
  weights = NA,
  savepath_tau = FALSE
)
}
\arguments{
\item{X}{\eqn{n} by \eqn{p} matrix of data.}

\item{epsilon}{tolerance for the convergence assessment.}

\item{AIC_eps}{if \code{AIC_selection=TRUE}, the convergence tolerance for the AIC convergence assessment.}

\item{theta}{initial value of the \eqn{p} by \eqn{p} precision matrix.}

\item{sigma}{initial value of the \eqn{p} by \eqn{p} covariance matrix.}

\item{Lambda_sq}{initial value of matrix of squared local shrinkage parameters.}

\item{tau_sq}{initial value of squared global shrinkage parameter. Not used if \code{AIC_selection=TRUE}.}

\item{tau_sq_min}{if \code{AIC_selection=TRUE}, the smallest value of \code{tau_sq} to consider.}

\item{tau_sq_stepsize}{if \code{AIC_selection=TRUE}, the step-size to use in the grid for \code{tau_sq}. Optional argument.}

\item{tau_sq_max}{if \code{AIC_selection=TRUE}, the largest value of \code{tau_sq} to consider.}

\item{stop_overflow}{should measures be taken to avoid overflow? Should be set to \code{TRUE} for networks expected to be dense.}

\item{maxitr}{maximum number of iterations.}

\item{verbose}{logical indicator of printing information at each iteration.}

\item{savepath}{logical indicator of saving the estimator at each iteration in the ECM algorithm. Only available for \eqn{p<200}.}

\item{save_Q}{should the value of the objective function at each step be saved?}

\item{fix_tau}{logical. Should \code{tau_sq} be fixed?}

\item{AIC_selection}{logical. Should the global shrinkage parameters be selected with AIC? Default \code{TRUE}.}

\item{stop_underflow}{should underflow be avoided by never allowing doubles to be smaller than the machine precision?}

\item{weights}{If provided, a vector of length \eqn{n} to weigh the observations with.}

\item{savepath_tau}{logical indicator of saving the value of \code{tau_sq} at each iteration in the ECM algorithm.}
}
\value{
a fitted \code{fastGHS} object
}
\description{
This function performs expectation-conditional-maximation for the graphical horseshoe
}
